pyproject.toml
[tool.poetry]
name = "learningbestcode"
version = "0.1.0"
description = "Поднимаю свои навыки написания чистого кода, использую FastAPI best practies, чистую архитектуру и различные другие подходы."
authors = ["Starobogatov Alexey <Starobogatov.a@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
#code lib
python = "^3.12"
fastapi = "^0.111.1"
aiogram = "^3.10.0"
pydantic = "^2.8.2"
pydantic-settings = "^2.3.4"
#data base lib
SQLAlchemy = {extras = ["asyncio"], version = "^2.0.31"}
redis = "^5.0.7"
asyncpg = "^0.29.0"
alembic = "^1.13.2"
#other
httpx = "^0.27.0"
uvicorn = "^0.30.3"
APScheduler = "^3.10.4"
cryptography = "^43.0.0"
bcrypt = "^4.2.0"
pyjwt = "^2.8.0"
black = "^24.4.2"
pytest-asyncio = "^0.23.8"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
lint.select = ["E", "F", "W", "C90"]
lint.ignore = ["E501"]
exclude = ["tests/*", "docs/*"]
